LayoutTests/inspector/console/console-big-array-expected.txt

CONSOLE MESSAGE: line 13: 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,100
CONSOLE MESSAGE: line 18: ,,,,,,,,,
CONSOLE MESSAGE: line 24: 0,1,2,3,4,5,6,7,8,9,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,100
CONSOLE MESSAGE: line 29: 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143,144,145,146,147,148,149,150,151,152,153,154,155,156,157,158,159,160,161,162,163,164,165,166,167,168,169,170,171,172,173,174,175,176,177,178,179,180,181,182,183,184,185,186,187,188,189,190,191,192,193,194,195,196,197,198,199,200,201,202,203,204,205,206,207,208,209,210,211,212,213,214,215,216,217,218,219,220,221,222,223,224,225,226,227,228,229,230,231,232,233,234,235,236,237,238,239,240,241,242,243,244,245,246,247,248,249,250,251,252,253,254,255,256,257,258,259,260,261,262,263,264,265,266,267,268,269,270,271,272,273,274,275,276,277,278,279,280,281,282,283,284,285,286,287,288,289,290,291,292,293,294,295,296,297,298,299,300,301,302,303,304,305,306,307,308,309,310,311,312,313,314,315,316,317,318,319,320,321,322,323,324,325,326,327,328,329,330,331,332,333,334,335,336,337,338,339,340,341,342,343,344,345,346,347,348,349,350,351,352,353,354,355,356,357,358,359,360,361,362,363,364,365,366,367,368,369,370,371,372,373,374,375,376,377,378,379,380,381,382,383,384,385,386,387,388,389,390,391,392,393,394,395,396,397,398,399,400,401,402,403,404
CONSOLE MESSAGE: line 42: %O
CONSOLE MESSAGE: line 52: %O
Tests that console logging dumps large arrays properly.



Array[101]
    [0 … 19]
    0: 0
    1: 1
    2: 2
    3: 3
    4: 4
    5: 5
    6: 6
    7: 7
    8: 8
    9: 9
    10: 10
    11: 11
    12: 12
    13: 13
    14: 14
    15: 15
    16: 16
    17: 17
    18: 18
    19: 19
    [20 … 39]
    20: 20
    21: 21
    22: 22
    23: 23
    24: 24
    25: 25
    26: 26
    27: 27
    28: 28
    29: 29
    30: 30
    31: 31
    32: 32
    33: 33
    34: 34
    35: 35
    36: 36
    37: 37
    38: 38
    39: 39
    [40 … 100]
    40: 40
    41: 41
    100: 100
    length: 101
    __proto__: Array[0] console-big-array.html:13

Array[10]
    0: undefined
    1: undefined
    2: undefined
    3: undefined
    4: undefined
    5: undefined
    6: undefined
    7: undefined
    8: undefined
    9: undefined
    length: 10
    __proto__: Array[0] console-big-array.html:18

Array[101]
    0: 0
    1: 1
    2: 2
    3: 3
    4: 4
    5: 5
    6: 6
    7: 7
    8: 8
    9: 9
    100: 100
    length: 101
    __proto__: Array[0] console-big-array.html:24

Array[405]
    [0 … 399]
    [0 … 19]
    [20 … 39]
    [40 … 59]
    [60 … 79]
    [80 … 99]
    [100 … 119]
    [120 … 139]
    [140 … 159]
    [160 … 179]
    [180 … 199]
    [200 … 219]
    [220 … 239]
    [240 … 259]
    [260 … 279]
    [280 … 299]
    [300 … 319]
    [320 … 339]
    [340 … 359]
    [360 … 379]
    [380 … 399]
    [400 … 404]
    400: 400
    401: 401
    402: 402
    403: 403
    404: 404
    length: 405
    __proto__: Array[0] console-big-array.html:29

Array[124]
    0: 0
    1: 1
    2: 2
    3: 3
    4: 4
    5: 5
    6: 6
    7: 7
    8: 8
    9: 9
    123: 123
    -Infinity: -Infinity
    3.14: 3.14
    4294967295: 4294967295
    4294967296: 4294967296
    Infinity: Infinity
    -123: -123
    NaN: NaN
    length: 124
    __proto__: Array[0] console-big-array.html:42

Array[4294967295]
    [0 … 19]
    0: 0
    1: 1
    2: 2
    3: 3
    4: 4
    5: 5
    6: 6
    7: 7
    8: 8
    9: 9
    10: 10
    11: 11
    12: 12
    13: 13
    14: 14
    15: 15
    16: 16
    17: 17
    18: 18
    19: 19
    [20 … 8388608]
    20: 20
    32: 32
    64: 64
    128: 128
    256: 256
    512: 512
    1024: 1024
    2048: 2048
    4096: 4096
    8192: 8192
    16384: 16384
    32768: 32768
    65536: 65536
    131072: 131072
    262144: 262144
    524288: 524288
    1048576: 1048576
    2097152: 2097152
    4194304: 4194304
    8388608: 8388608
    [16777216 … 4294967294]
    16777216: 16777216
    33554432: 33554432
    67108864: 67108864
    134217728: 134217728
    268435456: 268435456
    536870912: 536870912
    1073741824: 1073741824
    2147483648: 2147483648
    4294967294: 4294967294
    4294967296: 4294967296
    8589934592: 8589934592
    length: 4294967295
    __proto__: Array[0] console-big-array.html:52
